Move test to fsspec store
